Complete the LST changes to implement the new flow that were in progress. Remove files that are no longer used.  Cleanup memory leaks per Valgrind. Cleanup Python formatting per Pylint.
