+++ not-validated-prototype_lst/scripts/lst_utilities.py
@@ -152,15 +152,15 @@ class NARR(object):
 
     @staticmethod
     def dates(espa_metadata):
-        """Determines the before(t0), after(t1), and aquisition dates
+        """Determines the before(time_0), after(time_1), and aquisition dates
 
         Args:
             espa_metadata <espa.metadata>: The metadata for the data
 
         Returns:
-            acqusition <datetime>: Scene center date and time
-            t0 <datetime>: NARR data datetime before scene center
-            t1 <datetime>: NARR data datetime after scene center
+            acquisition <datetime>: Scene center date and time
+            time_0 <datetime>: NARR data datetime before scene center
+            time_1 <datetime>: NARR data datetime after scene center
         """
 
         center_time = str(espa_metadata.xml_object
@@ -172,8 +172,8 @@ class NARR(object):
         # Join them while dropping the last two '<number>Z'
         date_time = '-'.join([acq_date, center_time[:-2]])
 
-        acqusition = (datetime.datetime
-                      .strptime(date_time, '%Y-%m-%d-%H:%M:%S.%f'))
+        acquisition = (datetime.datetime
+                       .strptime(date_time, '%Y-%m-%d-%H:%M:%S.%f'))
 
         '''
         Determine the 3hr increments to use from the auxillary data
@@ -181,72 +181,26 @@ class NARR(object):
         and convert back to formatted strings
         '''
         scene_hour = int(center_time[:2])
-        t0_hour = scene_hour - (scene_hour % 3)
+        time_0_hour = scene_hour - (scene_hour % 3)
 
-        t0 = datetime.datetime(acqusition.year,
-                               acqusition.month,
-                               acqusition.day,
-                               t0_hour)
+        time_0 = datetime.datetime(acquisition.year,
+                                   acquisition.month,
+                                   acquisition.day,
+                                   time_0_hour)
 
-        t1 = t0 + datetime.timedelta(hours=3)
+        time_1 = time_0 + datetime.timedelta(hours=3)
 
-        return (acqusition, t0, t1)
+        # Round acquisition date to nearest minute.
+        if acquisition.second >= 30:
+            needed_seconds = datetime.timedelta(0, 60 - acquisition.second,
+                -acquisition.microsecond)
+            rounded_acquisition = acquisition + needed_seconds
+        else:
+            extra_seconds = datetime.timedelta(0, acquisition.second,
+                acquisition.microsecond)
+            rounded_acquisition = acquisition - extra_seconds
 
-
-class Metadata(object):
-    '''
-    Description:
-        Provides methods for interfacing with our metadata XML file.
-    '''
-
-    @staticmethod
-    def remove_products(xml_filename, product_list):
-        '''
-        Description:
-            Removes the specified products from the file system, as well as
-            from the XML file.
-        '''
-
-        if not product_list:
-            # We don't error, just nothing to do.
-            return
-
-#        espa_xml = metadata_api.parse(xml_filename, silence=True)
-        bands = espa_xml.get_bands()
-
-        # Gather all the filenames to be removed
-        filenames = list()
-        for band in bands.band:
-            if band.product in product_list:
-                # Add the .img file
-                filenames.append(band.file_name)
-                # Add the .hdr file
-                hdr_filename = band.file_name.replace('.img', '.hdr')
-                filenames.append(hdr_filename)
-
-        # If we found some then remove them
-        if len(filenames) > 0:
-            # First remove from disk
-            for filename in filenames:
-                if os.path.exists(filename):
-                    os.unlink(filename)
-
-            # Second remove from metadata XML
-            # Remove them from the XML by creating a new list of all the
-            # others
-            bands.band[:] = [band for band in bands.band
-                             if band.product not in product_list]
-
-#            try:
-#                # Export to the file with validation
-#                with open(xml_filename, 'w') as xml_fd:
-#                    metadata_api.export(xml_fd, espa_xml)
-#
-#            except Exception:
-#                raise
-
-        del bands
-        del espa_xml
+        return (rounded_acquisition, time_0, time_1)
 
 
 class Web(object):
@@ -330,104 +284,6 @@ class Web(object):
                     req.close()
 
         return status_code
-
-
-# ============================================================================
-# TODO TODO TODO - This Warp class is not usable yet, it is only a very very
-#                  early stage prototype.
-class Warp(object):
-    '''
-    Description:
-        Provides warping capabilities through GDAL's gdalwarp command line
-        tool.
-    '''
-
-    supported_output_formats = ['ENVI', 'GTiff']
-    x_pixel_min = 0.5
-    x_pixel_max = 1000.0
-    y_pixel_min = 0.5
-    y_pixel_max = 1000.0
-
-    def __init__(self):
-        self.base_cmd = ['gdalwarp', '-wm', '2048', '-multi']
-
-        self.x_pixel_size = None
-        self.y_pixel_size = None
-
-        self.source_proj4 = None
-        self.target_proj4 = None
-
-        self.overwrite_target = False
-
-        self.target_image_extents = None
-
-        self.output_format = None
-
-        self.source_no_data_value = None
-        self.target_no_data_value = None
-
-        self.source_files = None
-        self.target_file = None
-
-    def set_output_pixel_size(x_pixel_size, y_pixel_size):
-        if x_pixel_size < 0.5 or x_pixel_size > 1000.0:
-            raise Exception('X pixel_size out of range [{0} - {1}]'
-                            .format(self.x_pixel_min, self.x_pixel_max))
-
-        if y_pixel_size < 0.5 or y_pixel_size > 1000.0:
-            raise Exception('Y pixel_size out of range [{0} - {1}]'
-                            .format(self.y_pixel_min, self.y_pixel_max))
-
-        self.x_pixel_size = x_pixel_size
-        self.y_pixel_size = y_pixel_size
-
-    def set_source_proj4(source_proj4):
-        self.source_proj4 = source_proj4
-
-    def set_target_proj4(target_proj4):
-        self.target_proj4 = target_proj4
-
-    def set_source_no_data_value(no_data_value):
-        self.source_no_data_value = float(no_data_value)
-
-    def set_target_no_data_value(no_data_value):
-        self.target_no_data_value = float(no_data_value)
-
-    def overwrite_target(true_false):
-        self.overwrite_target = true_false
-
-    def set_output_format(format):
-
-        if format not in self.supported_output_formats:
-            raise NotImplementedError('Format [{0}] not supported'.
-                                      format(format))
-
-        self.output_format = format
-
-    def set_source_files(source_files):
-        self.source_files = source_files
-
-    def set_target_file(target_file):
-        self.target_file = target_file
-
-    def execute(self):
-        if self.source_files is None:
-            raise Exception('Source file(s) not specified')
-
-        if self.target_file is None:
-            raise Exception('Target file not specified')
-
-        # Add the base command
-        cmd = [x for x in self.base_cmd]
-
-        # Adde the pixel size command
-        if self.x_pixel_size is not None and self.y_pixel_size is not None:
-            cmd.extend(['-tr',
-                        str(self.x_pixel_size),
-                        str(self.y_pixel_size)])
-
-        # TODO TODO TODO - Lots to get done here
-        print ' '.join(cmd)
 
 
 class Geo(object):
