--- not-validated-prototype_lst/src/calculate_pixel_atmospheric_parameters.h
@@ -1,24 +1,0 @@
-
-#ifndef CALCULATE_PIXEL_ATMOSPHERIC_PARAMETERS_H
-#define CALCULATE_PIXEL_ATMOSPHERIC_PARAMETERS_H
-
-
-#include <stdbool.h>
-
-
-#include "lst_types.h"
-#include "input.h"
-
-
-int calculate_pixel_atmospheric_parameters
-(
-    Input_Data_t *input,       /* I: input structure */
-    REANALYSIS_POINTS *points, /* I: The coordinate points */
-    char *xml_filename,        /* I: XML filename */
-    double **modtran_results,  /* I: atmospheric parameter for MODTRAN run */
-    bool verbose               /* I: value to indicate if intermediate
-                                     messages will be printed */
-);
-
-
-#endif /* CALCULATE_PIXEL_ATMOSPHERIC_PARAMETERS_H */
