+++ not-validated-prototype_lst/src/output.c
@@ -33,8 +33,7 @@
       SUCCESS  No errors were encountered.
       ERROR    An error was encountered.
 ******************************************************************************/
-int
-add_lst_band_product
+int add_lst_band_product
 (
     char *xml_filename,
     char *reference_band_name,
